Extract superclass without context element
