+++ packages/ts/form/src/index.ts
@@ -1,6 +1,7 @@
 import type { VaadinWindow } from './types.js';
 
 export * from './Binder.js';
+export * from './BinderRoot.js';
 export * from './Field.js';
 export * from './Models.js';
 export * from './Validation.js';
